fix : 데이터 타입 변경

+++ src/main/java/live/ioteatime/frontservice/dto/OrganizationResponse.java
@@ -10,5 +10,5 @@ import java.math.BigInteger;
 public class OrganizationResponse {
     private int id;
     private String name;
-    private BigInteger electricityBudget;
+    private Long electricityBudget;
 }
